refactor: shared AsyncSqliteStore base for the edge + stats stores (PR6, task 971f8892)
StatsStore and EdgeStore carried a near-verbatim copy of the same ~150-line async-SQLite
connection-lifecycle machinery (#172): one WAL-mode autocommit connection reused across
ops, an op-lock serialising access, corrupt-file probe/quarantine on open, and
self-healing reconnect on a dead handle. Four methods were byte-identical; six more
differed only in a class name or db filename embedded in a log/assert string.

Extract it once into a new Foundation module, ``lithos.async_sqlite_store``. Both stores
now subclass ``AsyncSqliteStore`` and supply only what actually differs:

- ``SCHEMA`` (class attribute) and ``db_path`` (abstract property).
- StatsStore overrides ``_run_migrations`` (the three additive ALTERs it runs after the
  base schema); the base default is a no-op, so EdgeStore needs none.

The class-name / db-filename that made otherwise-identical bodies textually differ are
de-hardcoded via ``type(self).__name__`` / ``self.db_path.name`` — log output and backup
filenames are unchanged. Pure DRY, behaviour-neutral: the persistent-connection work
(#172) was already done in both stores, so nothing about the lifecycle semantics moves.

async_sqlite_store.py +272; edge_store.py 555→361; stats.py 1248→1027 — net -143
production lines. (The StatsStore god-store subdomain split — the other half of the task
description — is deliberately deferred; this PR is the shared base only.)

Architecture: the base is a new Foundation-tier component (SqliteStore), imported by
edge_store (Graph) and lcma.stats (LCMA). cross_component_edges 68→71 (+3: Graph→
SqliteStore, LCMA→SqliteStore, and SqliteStore→Config for the config-resolving db_path)
— an explicit, reviewed budget raise, the same shape as PR3's frontmatter_codec split;
the base's own component tells the truth about the shared dependency. import-linter holds
it to Foundation discipline (must not import Core/Entrypoints or the Core leaves).
lithos.edge_store added to [domain] exclude_modules: EdgeStore is flagged as a "model"
only because it now subclasses a local base (the scanner's local-base heuristic), but a
connection-lifecycle store is infrastructure, not a domain entity.

Behaviour-neutral: 115 store tests (test_edges + test_stats) pass unchanged as the oracle.
make check + make test-integration + make diagrams (48 guardrails) + lint-imports (3
contracts kept). All other metrics held (component_cycles 0, module_cycles 1,
modules_over_800 11, cross_module_private_refs 31).
