+++ docs/architecture.toml
@@ -50,6 +50,7 @@ Config          = "Pydantic-settings configuration and resolved store paths."
 Telemetry       = "OpenTelemetry tracing and metrics."
 Errors          = "Standard error types and canonical MCP error envelopes."
 Logging         = "Structured JSON logging setup."
+SqliteStore     = "Shared async-SQLite connection lifecycle (WAL, op-lock, corrupt-file probe/quarantine, self-healing reconnect) inherited by the edge and stats stores."
 
 [components]
 # lithos.pipeline is the composition root: it imports across Core to assemble the
@@ -87,13 +88,14 @@ Config          = ["lithos.config"]
 Telemetry       = ["lithos.telemetry"]
 Errors          = ["lithos.errors", "lithos.envelopes"]
 Logging         = ["lithos.logging_config"]
+SqliteStore     = ["lithos.async_sqlite_store"]
 
 # Tiers are for diagram grouping + documentation of the forbidden-import contract.
 # Dependencies must only point downward (Entrypoints -> Core -> Foundation).
 [tiers]
 Entrypoints = ["Entrypoints"]
 Core        = ["Knowledge", "Codec", "Search", "Graph", "Coordination", "Intake", "Provenance", "CognitiveMemory", "LCMA"]
-Foundation  = ["Events", "Config", "Telemetry", "Errors", "Logging"]
+Foundation  = ["Events", "Config", "Telemetry", "Errors", "Logging", "SqliteStore"]
 
 # Hard architecture budgets, checked by tests/guardrail/test_metrics_budgets.py
 # against the metrics snapshot (docs/generated/metrics.md explains each one).
@@ -107,7 +109,7 @@ Foundation  = ["Events", "Config", "Telemetry", "Errors", "Logging"]
 # is deliberately NOT pinned to the exact current high-water mark. Ratchet the
 # *count* of oversized modules down via modules_over_800_lines instead.
 [budgets]
-cross_component_edges  = 68    # measured 2026-07; task 4a3836a9 split lithos.frontmatter_codec into its own component (63→68) — no import changed; a component that imports BOTH the store and the format now counts as two edges where it was one. That split is what broke the Graph/Knowledge/Provenance/Search SCC (component_cycles 1→0); direction: down
+cross_component_edges  = 71    # measured 2026-07; task 971f8892 extracted lithos.async_sqlite_store as a shared Foundation base for the edge + stats stores (68→71): three new edges — edge_store (Graph)→SqliteStore, lcma.stats (LCMA)→SqliteStore, and SqliteStore→Config (the base owns the config-resolving db_path). Reviewed raise for a pure-DRY extraction — the base's own component tells the truth about the shared dependency (cf. frontmatter_codec 63→68); direction: down
 component_cycles       = 0    # measured 2026-07; task 4a3836a9 extracted the codec so graph.py/provenance.py import the pure format module, not KnowledgeManager — the 4-component Core SCC (Graph/Knowledge/Provenance/Search) is gone. PR1c had shrunk it from 7; goal: hold at 0
 module_cycles          = 1    # server<->tools; the knowledge<->graph<->search<->provenance SCC broke when the codec moved out (task 4a3836a9); goal: 0
 modules_over_800_lines = 11   # incl. tools.tasks (985) split from server in #374; knowledge.py 2444→1773 (codec out) but still over; direction: down
@@ -160,9 +162,12 @@ include_modules = [
     "lithos.search",
 ]
 # Not domain entities: config/settings, error types, LCMA-internal pipeline
-# working types (e.g. lcma.utils.Candidate), and the composition root's Pipeline
-# — a wiring container holding the built components, not a thing the domain has.
-exclude_modules = ["lithos.config", "lithos.errors", "lithos.lcma", "lithos.pipeline"]
+# working types (e.g. lcma.utils.Candidate), the composition root's Pipeline
+# (a wiring container, not a thing the domain has), and the SQLite stores —
+# EdgeStore is flagged only because it now subclasses the AsyncSqliteStore
+# base (a local-base heuristic), but a connection-lifecycle store is
+# infrastructure, not a domain model.
+exclude_modules = ["lithos.config", "lithos.errors", "lithos.lcma", "lithos.pipeline", "lithos.edge_store"]
 
 # Container / data-store view (docs/generated/containers.md). Store topology is
 # deployment knowledge, not derivable from imports — declared here, but every
